refactor: isolate leaf web server boundaries
